langchain/requirements.txt
streamlit
langchain
langchain-openai
langchain-classic
langchain_community
fastapi
uvicorn
openai
python-dotenv
tiktoken
faiss-cpu
pydantic

